Replace speedial with a buttongroup.
